# Make Style/HashSyntax a bit faster