The `RuboCop::ProcessedSource#valid_syntax?` check is rather slow, so let's first do a fast check against a simple regex. The regex will probably not match *all* correct inputs, in which case we fall back to calling `RuboCop::ProcessedSource#valid_syntax?`.

The regex was built by running RuboCop against the Rails 4.2.4 source code with a `puts sym_name` after the regex.

I have not written a performance test as such, but instead run it against a repository of ~2500 files:

    $ time rubocop -f o --only HashSyntax

    Before:
    rubocop -f o --only HashSyntax  22.83s user 0.30s system 99% cpu 23.147 total

    After:
    rubocop -f o --only HashSyntax  19.94s user 0.23s system 99% cpu 20.198 total