@rrosenblum 2015-08-31
Interesting change. Decent improvement in speed.

I have also been working on some performance/memory enhancements to RuboCop. I really need to put in some small pull requests.