@bbatsov 2015-09-03
I'm fine with the suggested change, but it has to be clarified a bit in the code and the changelog and this PR has to be rebased.